Add remark command with testings
RemarkCommand allows user to add remarks to a person in the
address book.

A new remark class is created to represent a remark. Each person
contains a remark object.This results in a change in how Person is
constructed. Hence, many files are updated to reflect this change.

Test file for RemarkCommand is also added.